Refactor transfers and stacks topics
Now we create three topics with three partitions each. The three
topics are created so that the three clickhouse servers can each
ingest data from one topic. Each topic has three partitions to make
sure that writing is balanced to the three Kafka brokers.

We have refactored the way migrations are done. To apply this change
on the kubernetes clusters we'll have to clean the Zookeeper migration
state.

+++ util/src/main/scala/net/santiment/util/MigrationUtil.scala
@@ -2,51 +2,93 @@ package net.santiment.util
 
 import java.util.concurrent.TimeUnit
 
+import com.typesafe.scalalogging.LazyLogging
 import org.apache.kafka.clients.admin.{AdminClient, NewTopic}
 
 import collection.JavaConverters._
 
-object MigrationUtil {
+object MigrationUtil extends LazyLogging {
+
+  def getTopics(kafka:AdminClient): java.util.Set[String] = {
+    kafka.listTopics().names().get()
+  }
 
   def topicMigration(kafka: AdminClient, topic:String, numPartitions:Int, replicationFactor:Short): Migration =
     Migration(
       name = s"Create topic $topic",
       up = ()=> {
-        val  config = Map[String,String](
-          ("compression.type", "lz4"),
-          //upper limit for fetch size
-          ("max.message.bytes", "52428800")
-        )
-        val t = new NewTopic(topic,numPartitions,replicationFactor)
+        if (getTopics(kafka).contains(topic)) {
+          logger.warn(s"Topic $topic already exists. Skipping creation")
+        } else {
 
-        t.configs(config.asJava)
-        kafka.createTopics(Seq(t).asJava)
+          val config = Map[String, String](
+            ("compression.type", "lz4"),
+            //upper limit for fetch size
+            ("max.message.bytes", "52428800")
+          )
+          val t = new NewTopic(topic, numPartitions, replicationFactor)
+
+          t.configs(config.asJava)
+          kafka.createTopics(Seq(t).asJava)
+        }
       },
 
       clean = ()=>{
-        kafka.deleteTopics(Seq(topic).asJava).all().get(30,TimeUnit.SECONDS)
+        if (getTopics(kafka).contains(topic)) {
+          kafka.deleteTopics(Seq(topic).asJava).all().get(30, TimeUnit.SECONDS)
+        } else {
+          logger.warn(s"Topic $topic does not exist. Skipping deletion")
+        }
       })
 
   def compactTopicMigration(kafka: AdminClient, topic:String, numPartitions:Int, replicationFactor:Short): Migration =
+    compactTopicMigration(kafka, Seq(topic), numPartitions, replicationFactor)
+
+  def compactTopicMigration(kafka: AdminClient, topics:Seq[String], numPartitions:Int, replicationFactor:Short): Migration =
     Migration(
-      name = s"Create topic $topic",
+      name = s"Create topics ${topics.mkString(",")}",
       up = ()=> {
-        val  config = Map[String,String](
-          ("compression.type", "lz4"),
-          ("cleanup.policy", "compact"),
-          //upper limit for fetch size
-          ("max.message.bytes", "52428800")
-        )
+        val oldTopics = getTopics(kafka)
 
-        val t = new NewTopic(topic,numPartitions,replicationFactor)
+        for ( topic <- topics if oldTopics.contains(topic) ) {
+          logger.warn(s"Topic $topic already exists. Skipping creation")
+        }
 
-        t.configs(config.asJava)
-        kafka.createTopics(Seq(t).asJava).all().get(30, TimeUnit.SECONDS)
+        val preparedTopics:Seq[NewTopic] = for (topic <- topics if !oldTopics.contains(topic) ) yield {
+
+            val config = Map[String, String](
+              ("compression.type", "lz4"),
+              ("cleanup.policy", "compact"),
+              //upper limit for fetch size
+              ("max.message.bytes", "52428800")
+            )
+
+            val t = new NewTopic(topic, numPartitions, replicationFactor)
+            t.configs(config.asJava)
+            t
+          }
+
+        kafka.createTopics(preparedTopics.asJava).all().get(30, TimeUnit.SECONDS)
+
       },
 
       clean = ()=>{
-        kafka.deleteTopics(Seq(topic).asJava)
+        val oldTopics = getTopics(kafka)
+
+        for ( topic <- topics if !oldTopics.contains(topic)) {
+          logger.warn(s"Topic $topic doesn't exist. Skipping deletion")
+        }
+
+        val preparedTopics: Seq[String] = topics.filter(oldTopics.contains(_))
+        kafka.deleteTopics(preparedTopics.asJava).all().get(30, TimeUnit.SECONDS)
       })
 
+
+  def cleanMigration(migration:Migration):Migration =
+    Migration(
+      name = s"rollback-${migration.name}",
+      up = migration.clean,
+      clean = ()=>{}
+    )
 
 }
